update post create router
